refactor(test): 중복되는 테스트 메서드 추출
