+++ src/main/java/wooteco/subway/line/dao/LineDao.java
@@ -1,4 +1,6 @@
-package wooteco.subway.line;
+package wooteco.subway.line.dao;
+
+import wooteco.subway.line.domain.Line;
 
 import java.util.List;
 import java.util.Optional;
