+++ src/main/java/wooteco/subway/line/dto/LineUpdateRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.line;
+package wooteco.subway.line.dto;
 
 public class LineUpdateRequest {
     private String name;
