+++ src/main/java/wooteco/subway/station/dao/MemoryStationDao.java
@@ -1,6 +1,7 @@
-package wooteco.subway.station;
+package wooteco.subway.station.dao;
 
 import org.springframework.util.ReflectionUtils;
+import wooteco.subway.station.domain.Station;
 
 import java.lang.reflect.Field;
 import java.util.ArrayList;
