+++ src/test/java/wooteco/subway/line/controller/LIneAcceptanceTest.java
@@ -1,22 +1,23 @@
-package wooteco.subway.line;
+package wooteco.subway.line.controller;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import wooteco.subway.AcceptanceTest;
+import wooteco.subway.line.dao.LineDao;
+import wooteco.subway.line.dto.LineRequest;
+import wooteco.subway.line.dto.LineResponse;
+import wooteco.subway.line.dto.LineSaveRequestDto;
 
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -34,47 +35,38 @@ public class LIneAcceptanceTest extends AcceptanceTest {
     @Test
     void createStation() throws JsonProcessingException {
         // given
-        LineRequest lineRequest = new LineRequest("신분당선", "bg-red-600");
+        LineSaveRequestDto lineRequest = new LineSaveRequestDto("신분당선", "bg-red-600");
         String content = objectMapper.writeValueAsString(lineRequest);
 
         // when
-        ExtractableResponse<Response> response = RestAssured.given().log().all()
-                .body(content)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post("/lines")
-                .then().log().all()
-                .extract();
+        ExtractableResponse<Response> response = addLine(content);
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
         assertThat(response.header("Location")).isNotBlank();
     }
 
-    @DisplayName("기존에 존재하는 노선역 이름으로 노선을 생성한다.")
-    @Test
-    void createStationWithDuplicateName() throws JsonProcessingException {
-        // given
-        LineRequest lineRequest = new LineRequest("신분당선", "bg-red-600");
-        String content = objectMapper.writeValueAsString(lineRequest);
-
-        RestAssured.given().log().all()
+    private ExtractableResponse<Response> addLine(String content) {
+        return RestAssured.given().log().all()
                 .body(content)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when()
                 .post("/lines")
                 .then().log().all()
                 .extract();
+    }
+
+    @DisplayName("기존에 존재하는 노선역 이름으로 노선을 생성한다.")
+    @Test
+    void createStationWithDuplicateName() throws JsonProcessingException {
+        // given
+        LineSaveRequestDto lineRequest = new LineSaveRequestDto("신분당선", "bg-red-600");
+        String content = objectMapper.writeValueAsString(lineRequest);
+
+        addLine(content);
 
         // when
-        ExtractableResponse<Response> response = RestAssured.given().log().all()
-                .body(content)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post("/lines")
-                .then()
-                .log().all()
-                .extract();
+        ExtractableResponse<Response> response = addLine(content);
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
@@ -87,24 +79,12 @@ public class LIneAcceptanceTest extends AcceptanceTest {
         LineRequest lineRequest1 = new LineRequest("신분당선", "bg-red-600");
         String content1 = objectMapper.writeValueAsString(lineRequest1);
 
-        ExtractableResponse<Response> createResponse1 = RestAssured.given().log().all()
-                .body(content1)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post("/lines")
-                .then().log().all()
-                .extract();
+        ExtractableResponse<Response> createResponse1 = addLine(content1);
 
         LineRequest lineRequest2 = new LineRequest("백기선", "bg-red-600");
         String content2 = objectMapper.writeValueAsString(lineRequest2);
 
-        ExtractableResponse<Response> createResponse2 = RestAssured.given().log().all()
-                .body(content2)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post("/lines")
-                .then().log().all()
-                .extract();
+        ExtractableResponse<Response> createResponse2 = addLine(content2);
 
         // when
         ExtractableResponse<Response> response = RestAssured.given().log().all()
@@ -136,13 +116,7 @@ public class LIneAcceptanceTest extends AcceptanceTest {
         String content = objectMapper.writeValueAsString(lineRequest);
 
         // when
-        ExtractableResponse<Response> response = RestAssured.given().log().all()
-                .body(content)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post("/lines")
-                .then().log().all()
-                .extract();
+        ExtractableResponse<Response> response = addLine(content);
 
         ExtractableResponse<Response> findLineResponse = RestAssured.given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
@@ -170,13 +144,7 @@ public class LIneAcceptanceTest extends AcceptanceTest {
         String content2 = objectMapper.writeValueAsString(lineRequest2);
 
         // when
-        ExtractableResponse<Response> response1 = RestAssured.given().log().all()
-                .body(content1)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post("/lines")
-                .then().log().all()
-                .extract();
+        ExtractableResponse<Response> response1 = addLine(content1);
 
         long newId = response1.body().jsonPath().getLong("id");
 
@@ -200,13 +168,7 @@ public class LIneAcceptanceTest extends AcceptanceTest {
         String content = objectMapper.writeValueAsString(lineRequest);
 
         // when
-        ExtractableResponse<Response> createResponse = RestAssured.given().log().all()
-                .body(content)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post("/lines")
-                .then().log().all()
-                .extract();
+        ExtractableResponse<Response> createResponse = addLine(content);
 
         long deleteId = createResponse.body().jsonPath().getLong("id");
         String uri = createResponse.header("Location");
