+++ src/test/java/wooteco/subway/line/dao/MemoryLineDaoTest.java
@@ -1,9 +1,12 @@
-package wooteco.subway.line;
+package wooteco.subway.line.dao;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.line.dao.LineDao;
+import wooteco.subway.line.dao.MemoryLineDao;
+import wooteco.subway.line.domain.Line;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
