+++ src/test/java/wooteco/subway/station/dao/MemoryStationDaoTest.java
@@ -1,8 +1,11 @@
-package wooteco.subway.station;
+package wooteco.subway.station.dao;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.station.dao.MemoryStationDao;
+import wooteco.subway.station.dao.StationDao;
+import wooteco.subway.station.domain.Station;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
