+++ src/test/java/wooteco/subway/station/domain/StationTest.java
@@ -1,27 +1,28 @@
-package wooteco.subway.line;
+package wooteco.subway.station.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import wooteco.subway.station.domain.Station;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class LineTest {
+class StationTest {
 
     @Test
-    @DisplayName("라인 정상 생성 테스트 ")
+    @DisplayName("역 객체를 생성한다.")
     void create() {
-        assertThatCode(() -> new Line("신분당선", "bg-red-600"))
+        assertThatCode(() -> new Station(1L, "아마찌역"))
                 .doesNotThrowAnyException();
     }
 
     @ParameterizedTest
-    @ValueSource(strings = {"아마찌", "아마찌선!"})
-    @DisplayName("잘 못된 이름의 라인이 들어올 시 예외가 발생한다")
+    @ValueSource(strings = {"아마찌", "아마찌역!"})
+    @DisplayName("잘 못된 이름의 역이 들어올 시 예외가 발생한다")
     void createException(String name) {
-        assertThatThrownBy(() -> new Line(1L, name, "bg-red-600"))
+        assertThatThrownBy(() -> new Station(1L, name))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 }
